services/analytics/requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
grpcio-health-checking==1.68.0
numpy>=1.26.0
pandas>=2.0.0
psycopg2-binary==2.9.10
APScheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.0
pydantic-settings==2.7.0
pytz==2024.2
